% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriously.R
\name{seriously}
\alias{seriously}
\title{Correct your previous console command}
\usage{
seriously(require_pkg = TRUE)
}
\arguments{
\item{require_pkg}{boolean, load the newly installed package or not.}
}
\value{
Nothing, use it for the side-effect.
}
\description{
This function guesses how to correct the error from your previous command during interactive usage.
Currently it can fix "There is no package called..." error.
}
